New run version

+++ lib.py
@@ -68,7 +68,7 @@ def is_prime(n):
     if n % 2 == 0:
         return False
 
-    for i in range(3, ceil(sqrt(n)) + 1, 2):
+    for i in range(3, int(ceil(sqrt(n))) + 1, 2):
         if n % i == 0:
             return False
